# Fix for exploding path counts