Updates the load test query to better reflect the current trade traversal query, which has significantly better query performance and returns far fewer paths.

## Test results

### Segment Date Filter

Arguments:

```
QUERY_RESPONSE_TYPE=graph_and_paths QUERY_DATE_FILTER=segment_date_filter COUNT=1
```

Before:

```
Success. Time 21301ms. Entity count 1376. Path count 576434 (TOTAL: 1 reqs 21 sec) (AVG: latency 21301ms, throughput 2.82 req/min, entities 1376) [Success 100% (1/1) Error 0% (0/1)] yOz94xUg8u04431rqikk2g
```

After:

```
Success. Time 5952ms. Entity count 1854. Path count 12245 (TOTAL: 1 reqs 6 sec) (AVG: latency 5953ms, throughput 10.08 req/min, entities 1854) [Success 100% (1/1) Error 0% (0/1)] yOz94xUg8u04431rqikk2g
```

### Path Date Filter

Arguments:

```
QUERY_RESPONSE_TYPE=graph_and_paths QUERY_DATE_FILTER=path_date_filter COUNT=1
```

Before:

```
Success. Time 31174ms. Entity count 6132. Path count 752992 (TOTAL: 1 reqs 31 sec) (AVG: latency 31175ms, throughput 1.92 req/min, entities 6132) [Success 100% (1/1) Error 0% (0/1)] yOz94xUg8u04431rqikk2g
```

After:

```
Success. Time 10451ms. Entity count 6412. Path count 67519 (TOTAL: 1 reqs 10 sec) (AVG: latency 10452ms, throughput 5.74 req/min, entities 6412) [Success 100% (1/1) Error 0% (0/1)] yOz94xUg8u04431rqikk2g
```